libmsmcomm: switch type of ref id to uint32_t

+++ libmsmcomm/src/internal.h
@@ -60,7 +60,7 @@ struct msmcomm_message
 {
 	uint8_t group_id;
 	uint16_t msg_id;
-	uint8_t ref_id;
+	uint32_t ref_id;
 	struct message_descriptor *descriptor;
 	void *payload;
 };
